pyproject.toml
[project]
name = "voxum"
version = "0.1.0"
description = "Voice meeting summarizer CLI - watches Google Drive for audio files, transcribes, summarizes, and delivers"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "typer>=0.9.0",
    "litellm>=1.0.0",
    "google-api-python-client>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
    "apscheduler>=3.10.0",
    "resend>=0.8.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
voxum = "voxum.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
